Update common.inc.js
The $.mobile.activePage.data('url') is more reliable than drupalgap.path, especially after a back button click. If you are on a full node, click back, then the drupalgap.path is wrong, still thinks you are on the node. $.mobile.activePage.data('url') has the correct path.

This is a fix for issue #64.
https://github.com/signalpoint/DrupalGap/issues/64