Add prefix/suffix for theme_view results
You can already add a _suffix and _prefix for most themes, but in a view case, I think it's relevant to add the same feature only when the view results is not empty. 

For example, lets say you want to add viewB on top of viewA, but only if viewA outputs something.