Add post Promise functions
Currently, there's a global postRender process, it works like a charm, but I think there's missing something. A postRender for Promises, like a callback.

To justify this proposition, here's a case :

Lets say we have an "AddToCalendar" link inside each views row. 
So when we call theme_view, theres nothing we can do to add an event on each links.
Global postRender is of course triggered before the view query. We could add a jQuery ".on()" function on body, but it's not really interesting since my solution here solve that problem.

What do you think?