Scroll top when app's div content is updated
When I'm switching from routes to routes, I notice that the scroll is kept for some reason. 
I propose here to force it to go "back to top" each time a new render is triggered. This is a simple suggestion, I just found it interesting and less frustrating as a mobile user. It could also be interesting to activate this feature in a setting.